Change: 各ユーザープロフィールに共通する処理を'concerns'で共通メソッド化
DRYなコードを目指し、'controllers/concerns'に'common.rb'を作成し、JobOfferer/SeekerのProfileのbefore_actionを共通化した。
