+++ app/controllers/job_offerers/profiles_controller.rb
@@ -1,5 +1,8 @@
 class JobOfferers::ProfilesController < ApplicationController
+  include Common
+
   before_action :authenticate_job_offerer!, except: %i[index show]
+  before_action :profile_exists_already, only: %i[new create]
   before_action :profile_require_correct_user, only: %i[edit update]
 
   def index
@@ -12,8 +15,6 @@ class JobOfferers::ProfilesController < ApplicationController
   end
 
   def new
-    return redirect_to root_url, alert: '操作が無効です' unless current_job_offerer.profile.nil?
-
     @profile = current_job_offerer.build_profile
   end
 
@@ -49,13 +50,5 @@ class JobOfferers::ProfilesController < ApplicationController
       :bio,
       :avatar
     )
-  end
-
-  def profile_require_correct_user
-    @profile = JobOffererProfile.find(params[:job_offerer_id])
-    return unless @profile.present?
-    return unless @profile.job_offerer != current_job_offerer
-
-    redirect_back(fallback_location: root_path, alert: '権限がありません')
   end
 end
